+++ quilt-sync/src-tauri/src/model.rs
@@ -487,6 +487,36 @@ pub async fn login(
     Ok(())
 }
 
+pub async fn login_oauth(
+    model: &impl QuiltModel,
+    host: &quilt::uri::Host,
+    params: quilt::auth::OAuthParams,
+) -> Result<(), Error> {
+    model
+        .get_quilt()
+        .lock()
+        .await
+        .get_remote()
+        .login_oauth(host, params)
+        .await?;
+    Ok(())
+}
+
+pub async fn get_or_register_client(
+    model: &impl QuiltModel,
+    host: &quilt::uri::Host,
+    redirect_uri: &str,
+) -> Result<String, Error> {
+    let client = model
+        .get_quilt()
+        .lock()
+        .await
+        .get_remote()
+        .get_or_register_client(host, redirect_uri)
+        .await?;
+    Ok(client.client_id)
+}
+
 #[cfg(test)]
 pub mod mocks {
     use super::*;